requirements.txt
streamlit>=1.41.1
python-docx>=1.1.2
pypdf>=5.2.0
pypdfium2>=5.5.0
pytesseract>=0.3.13
python-dotenv>=1.0.1
openai>=1.61.0
pytest>=8.3.4
